refactor(search applications): engine rename for create flyout and logic
