--- x-pack/plugins/enterprise_search/public/applications/applications/components/engines/create_engine_logic.test.ts
@@ -1,155 +1,0 @@
-/*
- * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
- * or more contributor license agreements. Licensed under the Elastic License
- * 2.0; you may not use this file except in compliance with the Elastic License
- * 2.0.
- */
-
-import { LogicMounter } from '../../../__mocks__/kea_logic';
-
-import { HttpError, Status } from '../../../../../common/types/api';
-
-import { KibanaLogic } from '../../../shared/kibana';
-import { CreateSearchApplicationApiLogic } from '../../api/search_applications/create_search_application_api_logic';
-
-import { SEARCH_APPLICATIONS_PATH } from '../../routes';
-
-import { CreateEngineLogic, CreateEngineLogicValues } from './create_engine_logic';
-
-const DEFAULT_VALUES: CreateEngineLogicValues = {
-  createDisabled: true,
-  createEngineError: undefined,
-  createEngineStatus: Status.IDLE,
-  engineName: '',
-  engineNameStatus: 'incomplete',
-  formDisabled: false,
-  indicesStatus: 'incomplete',
-  selectedIndices: [],
-};
-
-const VALID_ENGINE_NAME = 'unit-test-001';
-const INVALID_ENGINE_NAME = 'TEST';
-const VALID_INDICES_DATA = ['search-index-01'];
-
-describe('CreateEngineLogic', () => {
-  const { mount: apiLogicMount } = new LogicMounter(CreateSearchApplicationApiLogic);
-  const { mount } = new LogicMounter(CreateEngineLogic);
-
-  beforeEach(() => {
-    jest.clearAllMocks();
-    jest.useRealTimers();
-    apiLogicMount();
-    mount();
-  });
-
-  it('has expected defaults', () => {
-    expect(CreateEngineLogic.values).toEqual(DEFAULT_VALUES);
-  });
-
-  describe('listeners', () => {
-    it('createEngine makes expected request action with VALID_ENGINE_NAME', () => {
-      jest.spyOn(CreateEngineLogic.actions, 'createEngineRequest');
-
-      CreateEngineLogic.actions.setEngineName(VALID_ENGINE_NAME);
-      CreateEngineLogic.actions.setSelectedIndices(VALID_INDICES_DATA);
-
-      CreateEngineLogic.actions.createEngine();
-
-      expect(CreateEngineLogic.actions.createEngineRequest).toHaveBeenCalledTimes(1);
-      expect(CreateEngineLogic.actions.createEngineRequest).toHaveBeenCalledWith({
-        indices: ['search-index-01'],
-        name: VALID_ENGINE_NAME,
-      });
-    });
-
-    it('createEngine makes expected request action with INVALID_ENGINE_NAME', () => {
-      jest.spyOn(CreateEngineLogic.actions, 'createEngineRequest');
-
-      CreateEngineLogic.actions.setEngineName(INVALID_ENGINE_NAME);
-      CreateEngineLogic.actions.setSelectedIndices(VALID_INDICES_DATA);
-
-      CreateEngineLogic.actions.createEngine();
-
-      expect(CreateEngineLogic.actions.createEngineRequest).toHaveBeenCalledTimes(1);
-      expect(CreateEngineLogic.actions.createEngineRequest).toHaveBeenCalledWith({
-        indices: ['search-index-01'],
-        name: INVALID_ENGINE_NAME,
-      });
-    });
-    it('createEngine returns error when duplicate search application is created', () => {
-      const httpError: HttpError = {
-        body: {
-          error: 'search_application_already_exists',
-          message: 'Search application name already taken. Choose another name.',
-          statusCode: 409,
-        },
-        fetchOptions: {},
-        request: {},
-      } as HttpError;
-      CreateSearchApplicationApiLogic.actions.apiError(httpError);
-      expect(CreateEngineLogic.values.createEngineError).toEqual(httpError);
-    });
-
-    it('engineCreated is handled and is navigated to Search application list page', () => {
-      jest.spyOn(CreateEngineLogic.actions, 'fetchEngines');
-      jest
-        .spyOn(KibanaLogic.values, 'navigateToUrl')
-        .mockImplementationOnce(() => Promise.resolve());
-      CreateSearchApplicationApiLogic.actions.apiSuccess({
-        result: 'created',
-      });
-      expect(KibanaLogic.values.navigateToUrl).toHaveBeenCalledWith(SEARCH_APPLICATIONS_PATH);
-
-      expect(CreateEngineLogic.actions.fetchEngines).toHaveBeenCalledTimes(1);
-    });
-  });
-  describe('selectors', () => {
-    describe('engineNameStatus', () => {
-      it('returns incomplete with empty engine name', () => {
-        expect(CreateEngineLogic.values.engineNameStatus).toEqual('incomplete');
-      });
-      it('returns complete with valid engine name', () => {
-        CreateEngineLogic.actions.setEngineName(VALID_ENGINE_NAME);
-
-        expect(CreateEngineLogic.values.engineNameStatus).toEqual('complete');
-      });
-      it('returns complete with invalid engine name', () => {
-        CreateEngineLogic.actions.setEngineName(INVALID_ENGINE_NAME);
-        expect(CreateEngineLogic.values.engineNameStatus).toEqual('complete');
-      });
-    });
-    describe('indicesStatus', () => {
-      it('returns incomplete with 0 indices', () => {
-        expect(CreateEngineLogic.values.indicesStatus).toEqual('incomplete');
-      });
-      it('returns complete with at least one index', () => {
-        CreateEngineLogic.actions.setSelectedIndices(VALID_INDICES_DATA);
-        expect(CreateEngineLogic.values.indicesStatus).toEqual('complete');
-      });
-    });
-    describe('createDisabled', () => {
-      it('false with valid data', () => {
-        CreateEngineLogic.actions.setSelectedIndices(VALID_INDICES_DATA);
-        CreateEngineLogic.actions.setEngineName(VALID_ENGINE_NAME);
-
-        expect(CreateEngineLogic.values.createDisabled).toEqual(false);
-      });
-      it('false with invalid data', () => {
-        CreateEngineLogic.actions.setSelectedIndices(VALID_INDICES_DATA);
-        CreateEngineLogic.actions.setEngineName(INVALID_ENGINE_NAME);
-
-        expect(CreateEngineLogic.values.createDisabled).toEqual(false);
-      });
-    });
-    describe('formDisabled', () => {
-      it('returns true while create request in progress', () => {
-        CreateSearchApplicationApiLogic.actions.makeRequest({
-          indices: [VALID_INDICES_DATA[0]],
-          name: VALID_ENGINE_NAME,
-        });
-
-        expect(CreateEngineLogic.values.formDisabled).toEqual(true);
-      });
-    });
-  });
-});
