+++ x-pack/plugins/translations/translations/zh-CN.json
@@ -11710,7 +11710,7 @@
     "xpack.enterpriseSearch.content.engine.searchPreview.result.moreFieldsButton": "{count} {count, plural, other {更多字段}}",
     "xpack.enterpriseSearch.content.engine.searchPreview.resultsPerPage.option.label": "{value} {value, plural, other {结果}}",
     "xpack.enterpriseSearch.content.engineList.deleteEngine.successToast.title": "{searchApplicationName} 已删除",
-    "xpack.enterpriseSearch.content.engines.createEngine.headerSubTitle": "请浏览我们的 {enginesDocsLink} 了解详情！",
+    "xpack.enterpriseSearch.searchApplications.createSearchApplication.headerSubTitle": "请浏览我们的 {docsLink} 了解详情！",
     "xpack.enterpriseSearch.searchApplications.list.indicesFlyout.subTitle": "查看与 {engineName} 关联的索引",
     "xpack.enterpriseSearch.searchApplications.list.table.column.view.indices": "{indicesCount, number} {indicesCount, plural, other {索引}}",
     "xpack.enterpriseSearch.content.index.connector.syncRules.description": "添加同步规则以定制要从 {indexName} 同步哪些数据。默认包括所有内容，并根据按所列顺序显示的已配置同步规则集验证文档。",
@@ -12950,13 +12950,13 @@
     "xpack.enterpriseSearch.content.engineList.deleteEngineModal.delete.description": "删除搜索应用程序是不可逆操作。您的索引不会受到影响。",
     "xpack.enterpriseSearch.content.engineList.deleteEngineModal.title": "永久删除此搜索应用程序？",
     "xpack.enterpriseSearch.searchApplications.list.table.column.actions.deleteSearchApplicationLabel": "删除此搜索应用程序",
-    "xpack.enterpriseSearch.content.engines.createEngine.header.createError.title": "创建搜索应用程序时出错",
-    "xpack.enterpriseSearch.content.engines.createEngine.header.docsLink": "搜索应用程序文档",
-    "xpack.enterpriseSearch.content.engines.createEngine.headerTitle": "创建搜索应用程序",
-    "xpack.enterpriseSearch.content.engines.createEngine.nameEngine.placeholder": "搜索应用程序名称",
-    "xpack.enterpriseSearch.content.engines.createEngine.nameEngine.title": "命名搜索应用程序",
-    "xpack.enterpriseSearch.content.engines.createEngine.selectIndices.title": "选择索引",
-    "xpack.enterpriseSearch.content.engines.createEngine.submit": "创建",
+    "xpack.enterpriseSearch.searchApplications.createSearchApplication.header.createError.title": "创建搜索应用程序时出错",
+    "xpack.enterpriseSearch.searchApplications.createSearchApplication.header.docsLink": "搜索应用程序文档",
+    "xpack.enterpriseSearch.searchApplications.createSearchApplication.headerTitle": "创建搜索应用程序",
+    "xpack.enterpriseSearch.searchApplications.createSearchApplication.searchApplicationName.placeholder": "搜索应用程序名称",
+    "xpack.enterpriseSearch.searchApplications.createSearchApplication.searchApplicationName.title": "命名搜索应用程序",
+    "xpack.enterpriseSearch.searchApplications.createSearchApplication.selectIndices.title": "选择索引",
+    "xpack.enterpriseSearch.searchApplications.createSearchApplication.submit": "创建",
     "xpack.enterpriseSearch.content.engines.enginesList.empty.description": "下面我们指导您创建首个搜索应用程序。",
     "xpack.enterpriseSearch.content.engines.enginesList.empty.title": "创建您的首个搜索应用程序",
     "xpack.enterpriseSearch.content.engines.indices.addIndicesFlyout.updateError.title": "更新引擎时出错",
